disallow newlines, refactor some things
