+++ src/tot.Tests/tot.Tests.csproj
@@ -7,7 +7,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="approvaltests" Version="5.3.0" />
+    <Compile Remove="TestResults\**" />
+    <EmbeddedResource Remove="TestResults\**" />
+    <None Remove="TestResults\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="fluentassertions" Version="5.10.3" />
     <PackageReference Include="microsoft.data.analysis" Version="0.4.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
