+++ src/tot/Program.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.CommandLine;
 using System.CommandLine.Builder;
 using System.CommandLine.Invocation;
 using System.CommandLine.IO;
@@ -10,9 +8,8 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
-using Humanizer;
-using totlib;
 using TimeSpanParserUtil;
+using totlib;
 
 namespace tot
 {
@@ -23,134 +20,51 @@ namespace tot
             return await CommandLineParser.InvokeAsync(args);
         }
 
-        public static Parser CommandLineParser { get; } =
-            CreateCommandLineParser();
+        public static Parser CommandLineParser { get; } = tot.CommandLineParser.Create();
 
-        public static Parser CreateCommandLineParser(IDataAccessor dataAccessor = null)
+        internal static ParseArgument<DateTime> ParseTimeArgument(IDataAccessor dataAccessor)
         {
-            var pathOption = new Option<DirectoryInfo>(
-                "--path",
-                description: "The path containing the time series",
-                getDefaultValue: () => new DirectoryInfo(Directory.GetCurrentDirectory()));
+            return result =>
+            {
+                var token = result.Tokens.SingleOrDefault()?.Value;
 
-            var timeOption = new Option<DateTime>(
-                new [] { "-t" , "--time" },
-                description: "The time to record with the event", 
-                parseArgument: result =>
+                var now = (dataAccessor?.Clock ?? SystemClock.Instance).Now;
+
+                if (token == null)
                 {
-                    var token = result.Tokens.SingleOrDefault()?.Value;
+                    return now;
+                }
 
-                    var now = (dataAccessor?.Clock ?? SystemClock.Instance).Now;
-
-                    if (token == null)
+                if (DateTime.TryParse(token, provider: null, styles: DateTimeStyles.NoCurrentDateDefault, out var specified))
+                {
+                    if (specified.Date == default)
                     {
-                        return now;
-                    }
-
-                    if (DateTime.TryParse(token, provider: null, styles: DateTimeStyles.NoCurrentDateDefault, out var specified))
-                    {
-                        if (specified.Date == default)
+                        if (now.TimeOfDay < specified.TimeOfDay)
                         {
-                            if (now.TimeOfDay < specified.TimeOfDay)
-                            {
-                                var yesterday = now.Subtract(TimeSpan.FromDays(1));
-                                return yesterday.Date.Add(specified.TimeOfDay);
-                            }
-                            else
-                            {
-                                return now.Date.Add(specified.TimeOfDay);
-                            }
+                            var yesterday = now.Subtract(TimeSpan.FromDays(1));
+                            return yesterday.Date.Add(specified.TimeOfDay);
                         }
-
-                        return specified;
+                        else
+                        {
+                            return now.Date.Add(specified.TimeOfDay);
+                        }
                     }
 
-                    if (TimeSpanParser.TryParse(token, out var timespan))
-                    {
-                        return now.Add(timespan);
-                    }
+                    return specified;
+                }
 
-                    result.ErrorMessage = $"Couldn't figure out what time \"{token}\" refers to.";
-
-                    return default;
-                }, isDefault: true);
-
-            var rootCommand = new RootCommand("tot")
-            {
-                Add(),
-                List(),
-                timeOption,
-                new Argument<string>("series").AddSuggestions((result, match) =>
+                if (TimeSpanParser.TryParse(token, out var timespan))
                 {
-                    if (result == null)
-                    {
-                        return Array.Empty<string>();
-                    }
+                    return now.Add(timespan);
+                }
 
-                    var path = result.ValueForOption(pathOption);
+                result.ErrorMessage = $"Couldn't figure out what time \"{token}\" refers to.";
 
-                    EnsureDataAccessorIsInitialized(path, ref dataAccessor);
-
-                    return dataAccessor.ListSeries();
-                }),
-                new Argument<IEnumerable<string>>("values")
+                return default;
             };
-
-            rootCommand.AddGlobalOption(pathOption);
-
-            rootCommand.Handler = CommandHandler.Create<string, string[], DateTime, DirectoryInfo, IConsole>((series, values, time, path, console) =>
-            {
-                EnsureDataAccessorIsInitialized(path, ref dataAccessor);
-
-                dataAccessor.AppendValues(series, time, values);
-
-                console.Out.WriteLine($"{time.Humanize(utcDate: false)}: {series} {string.Join(" ", values ?? Array.Empty<string>())}");
-            });
-
-            Command Add()
-            {
-                var command = new Command("add", "Adds a new series")
-                {
-                    new Argument<string>("name"),
-                    new Argument<IEnumerable<string>>("columns")
-                };
-
-                command.Handler = CommandHandler.Create<string, string[], DirectoryInfo>(async (name, columns, path) =>
-                {
-                    EnsureDataAccessorIsInitialized(path, ref dataAccessor);
-
-                    dataAccessor.CreateSeries(name, columns);
-                });
-
-                return command;
-            }
-
-            Command List()
-            {
-                var command = new Command("list", "Lists the defined series");
-
-                command.Handler = CommandHandler.Create<DirectoryInfo, IConsole>((path, console) =>
-                {
-                    EnsureDataAccessorIsInitialized(path, ref dataAccessor);
-
-                    var series = dataAccessor.ListSeries().OrderBy(s => s);
-
-                    console.Out.Write(
-                        string.Join(
-                            Environment.NewLine, series));
-                });
-
-                return command;
-            }
-
-            var builder = new CommandLineBuilder(rootCommand)
-                          .HandleExceptionsGracefully()
-                          .UseDefaults();
-
-            return builder.Build();
         }
 
-        public static CommandLineBuilder HandleExceptionsGracefully(this CommandLineBuilder builder) =>
+        public static CommandLineBuilder DisplayException(this CommandLineBuilder builder) =>
             builder.UseMiddleware(async (context, next) =>
             {
                 try
@@ -159,17 +73,23 @@ namespace tot
                 }
                 catch (TotException e)
                 {
+                    Console.ForegroundColor = ConsoleColor.Red;
                     context.Console.Error.WriteLine(e.Message);
                     context.ResultCode = 1;
                 }
                 catch (TargetInvocationException e) when (e.InnerException is TotException)
                 {
+                    Console.ForegroundColor = ConsoleColor.Red;
                     context.Console.Error.WriteLine(e.InnerException.Message);
                     context.ResultCode = 1;
                 }
+                finally
+                {
+                    Console.ResetColor();
+                }
             }, MiddlewareOrder.ExceptionHandler);
 
-        private static void EnsureDataAccessorIsInitialized(
+        internal static void EnsureDataAccessorIsInitialized(
             DirectoryInfo path,
             ref IDataAccessor dataAccessor) =>
             dataAccessor ??=
