+++ src/totlib/IDataAccessor.cs
@@ -15,23 +15,4 @@ namespace totlib
 
         IClock Clock { get; }
     }
-
-    public static class DataAccessor
-    {
-        public static void AppendValues(
-            this IDataAccessor dataAccessor,
-            string seriesName,
-            params string[] values)
-        {
-            dataAccessor.AppendValues(seriesName, default,  values);
-        }
-
-        public static void CreateSeries(
-            this IDataAccessor dataAccessor,
-            string seriesName,
-            params string[] columnNames)
-        {
-            dataAccessor.CreateSeries(seriesName, columnNames);
-        }
-    }
 }
